Move `fromJson` and `toJson` functions of domain types to extension functions in 'infrastructure'.
As part of this, all tests depending on serialization were also moved to an 'infrastructure' namespace.

There are still serialization dependencies left in the domain model. Namely, 'Custom...' types and custom serializers (e.g., `DeviceDescriptorSerializer`).
