+++ carp.deployment.core/src/commonTest/kotlin/dk/cachet/carp/deployment/domain/StudyDeploymentTest.kt
@@ -1,9 +1,9 @@
 package dk.cachet.carp.deployment.domain
 
 import dk.cachet.carp.common.UUID
-import dk.cachet.carp.common.serialization.JSON
-import dk.cachet.carp.protocols.domain.*
+import dk.cachet.carp.common.serialization.createDefaultJSON
 import dk.cachet.carp.protocols.domain.devices.*
+import kotlinx.serialization.json.Json
 import kotlin.test.*
 
 
@@ -12,6 +12,12 @@ import kotlin.test.*
  */
 class StudyDeploymentTest
 {
+    companion object
+    {
+        private val JSON: Json = createDefaultJSON()
+    }
+
+
     @Test
     fun cant_initialize_deployment_with_errors()
     {
@@ -22,28 +28,6 @@ class StudyDeploymentTest
         assertFailsWith<IllegalArgumentException>
         {
             StudyDeployment( snapshot, testId )
-        }
-    }
-
-    @Test
-    fun cant_initialize_deployment_with_invalid_snapshot()
-    {
-        // Initialize valid protocol.
-        val protocol = createEmptyProtocol()
-        val master = StubMasterDeviceDescriptor( "Master" )
-        val connected = StubMasterDeviceDescriptor( "Connected" )
-        protocol.addMasterDevice( master )
-        protocol.addConnectedDevice( connected, master )
-        val snapshot = protocol.getSnapshot()
-
-        // Create invalid snapshot by editing JSON.
-        val json = snapshot.toJson()
-        val invalidJson = json.replaceFirst( "\"Master\"", "\"Non-existing device\"" )
-        val invalidSnapshot = StudyProtocolSnapshot.fromJson( invalidJson )
-
-        assertFailsWith<IllegalArgumentException>
-        {
-            StudyDeployment( invalidSnapshot, testId )
         }
     }
 
@@ -191,23 +175,6 @@ class StudyDeploymentTest
         assertEquals(
             deployment.registeredDevices.count(),
             deployment.registeredDevices.entries.intersect( fromSnapshot.registeredDevices.entries ).count() )
-    }
-
-    @Test
-    fun create_deployment_fromSnapshot_with_custom_extending_types_succeeds()
-    {
-        val protocol = createEmptyProtocol()
-        val master = UnknownMasterDeviceDescriptor( "Unknown" )
-        protocol.addMasterDevice( master )
-        val deployment = deploymentFor( protocol )
-        deployment.registerDevice( master, UnknownDeviceRegistration( "0" ) )
-
-        var serialized: String = deployment.getSnapshot().toJson()
-        serialized = serialized.replace( "dk.cachet.carp.deployment.domain.UnknownMasterDeviceDescriptor", "com.unknown.CustomMasterDevice" )
-        serialized = serialized.replace( "dk.cachet.carp.deployment.domain.UnknownDeviceRegistration", "com.unknown.CustomDeviceRegistration" )
-
-        val snapshot = DeploymentSnapshot.fromJson( serialized )
-        StudyDeployment.fromSnapshot( snapshot )
     }
 
     @Test
