+++ carp.protocols.core/src/commonMain/kotlin/dk/cachet/carp/protocols/domain/devices/CustomDeviceRegistration.kt
@@ -10,6 +10,12 @@ import kotlinx.serialization.json.*
 data class CustomDeviceRegistration( override val className: String, override val jsonSource: String )
     : DeviceRegistration(), UnknownPolymorphicWrapper
 {
+    companion object
+    {
+        private val JSON: Json = createDefaultJSON()
+    }
+
+
     override val deviceId: String
 
     init
