+++ carp.protocols.core/src/commonTest/kotlin/dk/cachet/carp/protocols/domain/CreateTestObjects.kt
@@ -143,44 +143,3 @@ internal data class UnknownTrigger( override val sourceDeviceRoleName: String ) 
     }
 }
 
-/**
- * Creates a study protocol which includes:
- * (1) an unknown master device and unknown connected device
- * (2) unknown task with an unknown measure and unknown data type, triggered by an unknown trigger
- * (3) known task with an unknown measure and known data type
- * There is thus exactly one unknown object for each of these types, except for 'Measure' which has two.
- */
-fun serializeProtocolSnapshotIncludingUnknownTypes(): String
-{
-    val protocol = createComplexProtocol()
-
-    // (1) Add unknown master with unknown connected device.
-    val master = UnknownMasterDeviceDescriptor( "Unknown" )
-    protocol.addMasterDevice( master )
-    val connected = UnknownDeviceDescriptor( "Unknown 2" )
-    protocol.addConnectedDevice( connected, master )
-
-    // (2) Add unknown task with unknown measure.
-    val measures: List<Measure> = listOf( UnknownMeasure( STUB_DATA_TYPE ), StubMeasure( STUB_DATA_TYPE ) )
-    val task = UnknownTaskDescriptor( "Unknown task", measures )
-    val trigger = UnknownTrigger( master.roleName )
-    protocol.addTriggeredTask( trigger, task, master )
-
-    // (3) Add known task with unknown measure.
-    val task2 = StubTaskDescriptor( "Known task", listOf( UnknownMeasure( STUB_DATA_TYPE ) ) )
-    protocol.addTriggeredTask( trigger, task2, master )
-
-    val snapshot: StudyProtocolSnapshot = protocol.getSnapshot()
-    var serialized: String = snapshot.toJson()
-
-    // Replace the strings which identify the types to load by the PolymorphicSerializer.
-    // This will cause the types not to be found while deserializing, hence mimicking 'custom' types.
-    serialized = serialized.replace( "dk.cachet.carp.protocols.domain.UnknownMasterDeviceDescriptor", "com.unknown.CustomMasterDevice" )
-    serialized = serialized.replace( "dk.cachet.carp.protocols.domain.UnknownDeviceDescriptor", "com.unknown.CustomDevice" )
-    serialized = serialized.replace( "dk.cachet.carp.protocols.domain.UnknownTaskDescriptor", "com.unknown.CustomTask" )
-    serialized = serialized.replace( "dk.cachet.carp.protocols.domain.UnknownMeasure", "com.unknown.CustomMeasure" )
-    serialized = serialized.replace( "dk.cachet.carp.protocols.domain.UnknownTrigger", "com.unknown.CustomTrigger" )
-
-    return serialized
-}
-
