+++ carp.protocols.core/src/commonTest/kotlin/dk/cachet/carp/protocols/domain/StudyProtocolSnapshotTest.kt
@@ -3,7 +3,6 @@ package dk.cachet.carp.protocols.domain
 import dk.cachet.carp.common.UUID
 import dk.cachet.carp.protocols.domain.devices.*
 import dk.cachet.carp.protocols.domain.tasks.*
-import dk.cachet.carp.protocols.domain.tasks.measures.CustomMeasure
 import dk.cachet.carp.protocols.domain.triggers.*
 import kotlin.test.*
 
@@ -63,64 +62,6 @@ class StudyProtocolSnapshotTest
         val newTriggeredTaskSnapshot = protocol.getSnapshot()
         assertTrue( newTaskSnapshot != newTriggeredTaskSnapshot )
         assertTrue( newTaskSnapshot.hashCode() != newTriggeredTaskSnapshot.hashCode() )
-    }
-
-    @Test
-    fun can_serialize_and_deserialize_snapshot_using_JSON()
-    {
-        val protocol: StudyProtocol = createComplexProtocol()
-        val snapshot: StudyProtocolSnapshot = protocol.getSnapshot()
-
-        val serialized: String = snapshot.toJson()
-        val parsed: StudyProtocolSnapshot = StudyProtocolSnapshot.fromJson( serialized )
-
-        assertEquals( snapshot, parsed )
-    }
-
-    /**
-     * Types not known at compile time should not prevent deserializing a protocol, but should be loaded through a 'Custom' type wrapper.
-     */
-    @Test
-    fun unknown_types_are_wrapped_when_deserializing()
-    {
-        val serialized: String = serializeProtocolSnapshotIncludingUnknownTypes()
-
-        val parsed = StudyProtocolSnapshot.fromJson( serialized )
-        assertEquals( 1, parsed.masterDevices.filter { m -> m is CustomMasterDeviceDescriptor }.count() )
-        assertEquals( 1, parsed.connectedDevices.filter { m -> m is CustomDeviceDescriptor }.count() )
-        assertEquals( 1, parsed.tasks.filter { m -> m is CustomTaskDescriptor }.count() )
-        val allMeasures = parsed.tasks.flatMap{ t -> t.measures }
-        assertEquals( 2, allMeasures.filter { m -> m is CustomMeasure }.count() )
-        assertEquals( 1, parsed.triggers.filter { t -> t.trigger is CustomTrigger }.count() )
-    }
-
-    @Test
-    fun unknown_connected_master_device_is_deserialized_as_a_master_device()
-    {
-        val protocol = createEmptyProtocol()
-        val master = StubMasterDeviceDescriptor( "Master" )
-        protocol.addMasterDevice( master )
-        val unknownMaster = UnknownMasterDeviceDescriptor( "Unknown master" )
-        protocol.addConnectedDevice( unknownMaster, master )
-
-        var serialized = protocol.getSnapshot().toJson()
-        serialized = serialized.replace( "dk.cachet.carp.protocols.domain.UnknownMasterDeviceDescriptor", "com.unknown.CustomMasterDevice" )
-
-        val parsed = StudyProtocolSnapshot.fromJson( serialized )
-        assertTrue { parsed.connectedDevices.single() is MasterDeviceDescriptor }
-    }
-
-    /**
-     * Types which were wrapped in a 'Custom' type wrapper upon deserialization should be serialized to their original form (returning the original type, not the wrapper).
-     */
-    @Test
-    fun serializing_unknown_types_removes_the_wrapper()
-    {
-        val serialized: String = serializeProtocolSnapshotIncludingUnknownTypes()
-        val snapshot = StudyProtocolSnapshot.fromJson( serialized )
-
-        val customSerialized = snapshot.toJson()
-        assertEquals( serialized, customSerialized )
     }
 
     @Test
