+++ carp.protocols.core/src/commonTest/kotlin/dk/cachet/carp/protocols/domain/devices/CustomMasterDeviceDescriptorTest.kt
@@ -1,8 +1,9 @@
 package dk.cachet.carp.protocols.domain.devices
 
-import dk.cachet.carp.common.serialization.JSON
+import dk.cachet.carp.common.serialization.createDefaultJSON
 import dk.cachet.carp.protocols.domain.*
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
 import kotlin.test.*
 
 
@@ -11,6 +12,12 @@ import kotlin.test.*
  */
 class CustomMasterDeviceDescriptorTest
 {
+    companion object
+    {
+        private val JSON: Json = createDefaultJSON()
+    }
+
+
     @Test
     fun initialization_from_json_extracts_base_MasterDeviceDescriptor_properties()
     {
